add precaution in case the install file is run accidentally

+++ install.sh
@@ -14,36 +14,41 @@ else
     exit 1
 fi
 
-timedatectl set-ntp true
+read -p "This is the installation script with low-level modifications. Are you sure you want to continue? " -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]; then
 
-# partitioning and mounting disk
-echo DISK="$DISK", PARTITION="$PARTITION"
+    timedatectl set-ntp true
 
-parted -s "$DISK" mklabel msdos
-parted -s -a optimal "$DISK" mkpart primary ext4 0% 100%
-parted -s "$DISK" set 1 boot on
-mkfs.ext4 -F "$PARTITION"
+    # partitioning and mounting disk
+    echo DISK="$DISK", PARTITION="$PARTITION"
 
-mount "$PARTITION" /mnt
+    parted -s "$DISK" mklabel msdos
+    parted -s -a optimal "$DISK" mkpart primary ext4 0% 100%
+    parted -s "$DISK" set 1 boot on
+    mkfs.ext4 -F "$PARTITION"
 
-# installation
-echo 'Server = http://mirror.internode.on.net/pub/archlinux/$repo/os/$arch' > /etc/pacman.d/mirrorlist
-pacman -Syy
+    mount "$PARTITION" /mnt
 
-pacstrap /mnt base base-devel emacs grub ntp openssh sudo wget
-genfstab -U /mnt >> /mnt/etc/fstab
+    # installation
+    echo 'Server = http://mirror.internode.on.net/pub/archlinux/$repo/os/$arch' > /etc/pacman.d/mirrorlist
+    pacman -Syy
 
-wget https://raw.githubusercontent.com/roxolan/vk-arch/master/chroot.sh
-chmod +x ./chroot.sh
-cp -p ./chroot.sh /mnt
-# cp ~/.ssh/authorized_keys /mnt
+    pacstrap /mnt base base-devel emacs grub ntp openssh sudo wget
+    genfstab -U /mnt >> /mnt/etc/fstab
 
-arch-chroot /mnt ./chroot.sh "$DISK"
-rm /mnt/chroot.sh
-# rm /mnt/authorized_keys
+    wget https://raw.githubusercontent.com/roxolan/vk-arch/master/chroot.sh
+    chmod +x ./chroot.sh
+    cp -p ./chroot.sh /mnt
+    # cp ~/.ssh/authorized_keys /mnt
 
-umount -R /mnt
-systemctl reboot
+    arch-chroot /mnt ./chroot.sh "$DISK"
+    rm /mnt/chroot.sh
+    # rm /mnt/authorized_keys
 
-# installing oh-my-zsh is postponed until git is present in the system
-# sh -c "$(wget https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O -)"
+    umount -R /mnt
+    systemctl reboot
+
+    # installing oh-my-zsh is postponed until git is present in the system
+    # sh -c "$(wget https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O -)"
+
+fi
